+++ media/jni/android_media_MediaExtractor.cpp
@@ -84,7 +84,10 @@ class JavaDataSourceBridge : public DataSource {
         jbyteArray byteArrayObj = env->NewByteArray(size);
         env->DeleteLocalRef(env->GetObjectClass(mDataSource));
         env->DeleteLocalRef(env->GetObjectClass(byteArrayObj));
-        ssize_t numread = env->CallIntMethod(mDataSource, mReadMethod, offset, byteArrayObj, size);
+        ssize_t numread = env->CallIntMethod(mDataSource, mReadMethod,
+                                             static_cast<jlong>(offset),
+                                             byteArrayObj,
+                                             static_cast<jint>(size));
         env->GetByteArrayRegion(byteArrayObj, 0, size, (jbyte*) buffer);
         env->DeleteLocalRef(byteArrayObj);
         if (env->ExceptionCheck()) {
@@ -627,12 +630,12 @@ static jboolean android_media_MediaExtractor_getSampleCryptoInfo(
     env->CallVoidMethod(
             cryptoInfoObj,
             gFields.cryptoInfoSetID,
-            numSubSamples,
+            static_cast<jint>(numSubSamples),
             numBytesOfPlainDataObj,
             numBytesOfEncryptedDataObj,
             keyObj,
             ivObj,
-            mode);
+            static_cast<jint>(mode));
 
     return JNI_TRUE;
 }
